+++ projects/small_fem/formulation/TermFieldField.h
@@ -26,7 +26,7 @@ class TermFieldField: public Term{
                 fullMatrix<double>**& cM);
 
   void computeB(const GroupOfJacobian& goj,
-                unsigned int nG,
+                size_t nG,
                 fullMatrix<double>**& bM);
 };
